new AuthorGrid component + update ContentSummary
- renamed ContentSummary to ContentCard and changed component to render
  the content instead;
